Sort of removed platform-specific crap from test005. This changes the simple dynamic loader interface in that the file extension (.so or .dll) must NOT be specified now. The platform-specific code will append the proper extension.
Also, the path to the test component is no longer specified in the test. This works under Windows because the LoadLibrary API looks in many different places for the .dll file, including the directory from which the calling process' exe was loaded (so it's not a coincidence that the testcomponent project now outputs the dll in test005/).